+++ pydis_core/_bot.py
@@ -9,11 +9,11 @@ import aiohttp
 import discord
 from discord.ext import commands
 
-from botcore.async_stats import AsyncStatsClient
-from botcore.site_api import APIClient
-from botcore.utils import scheduling
-from botcore.utils._extensions import walk_extensions
-from botcore.utils.logging import get_logger
+from pydis_core.async_stats import AsyncStatsClient
+from pydis_core.site_api import APIClient
+from pydis_core.utils import scheduling
+from pydis_core.utils._extensions import walk_extensions
+from pydis_core.utils.logging import get_logger
 
 try:
     from async_rediscache import RedisSession
@@ -55,7 +55,7 @@ class BotBase(commands.Bot):
             allowed_roles: A list of role IDs that the bot is allowed to mention.
             http_session (aiohttp.ClientSession): The session to use for the bot.
             redis_session: The `async_rediscache.RedisSession`_ to use for the bot.
-            api_client: The :obj:`botcore.site_api.APIClient` instance to use for the bot.
+            api_client: The :obj:`pydis_core.site_api.APIClient` instance to use for the bot.
             statsd_url: The URL of the statsd server to use for the bot. If not given,
                 a dummy statsd client will be created.
 
@@ -73,7 +73,7 @@ class BotBase(commands.Bot):
         self.statsd_url = statsd_url
 
         if redis_session and not REDIS_AVAILABLE:
-            warnings.warn("redis_session kwarg passed, but async-rediscache not installed!")
+            warnings.warn("redis_session kwarg passed, but async-rediscache not installed!", stacklevel=2)
         elif redis_session:
             self.redis_session = redis_session
 
@@ -82,6 +82,7 @@ class BotBase(commands.Bot):
 
         self._statsd_timerhandle: Optional[asyncio.TimerHandle] = None
         self._guild_available: Optional[asyncio.Event] = None
+        self._extension_loading_task: asyncio.Task | None = None
 
         self.stats: Optional[AsyncStatsClient] = None
 
@@ -116,17 +117,30 @@ class BotBase(commands.Bot):
                 attempt + 1
             )
 
-    async def load_extensions(self, module: types.ModuleType) -> None:
-        """
-        Load all the extensions within the given module and save them to ``self.all_extensions``.
-
-        This should be ran in a task on the event loop to avoid deadlocks caused by ``wait_for`` calls.
-        """
+    async def _load_extensions(self, module: types.ModuleType) -> None:
+        """Load all the extensions within the given module and save them to ``self.all_extensions``."""
         await self.wait_until_guild_available()
         self.all_extensions = walk_extensions(module)
 
         for extension in self.all_extensions:
             scheduling.create_task(self.load_extension(extension))
+
+    async def _sync_app_commands(self) -> None:
+        """Sync global & guild specific application commands after extensions are loaded."""
+        await self._extension_loading_task
+        await self.tree.sync()
+        await self.tree.sync(guild=discord.Object(self.guild_id))
+
+    async def load_extensions(self, module: types.ModuleType, sync_app_commands: bool = True) -> None:
+        """
+        Load all the extensions within the given ``module`` and save them to ``self.all_extensions``.
+
+        Args:
+            sync_app_commands: Whether to sync app commands after all extensions are loaded.
+        """
+        self._extension_loading_task = scheduling.create_task(self._load_extensions(module))
+        if sync_app_commands:
+            scheduling.create_task(self._sync_app_commands())
 
     def _add_root_aliases(self, command: commands.Command) -> None:
         """Recursively add root aliases for ``command`` and any of its subcommands."""
@@ -220,7 +234,7 @@ class BotBase(commands.Bot):
         An async init to startup generic services.
 
         Connects to statsd, and calls
-        :func:`AsyncStatsClient.create_socket <botcore.async_stats.AsyncStatsClient.create_socket>`
+        :func:`AsyncStatsClient.create_socket <pydis_core.async_stats.AsyncStatsClient.create_socket>`
         and :func:`ping_services`.
         """
         loop = asyncio.get_running_loop()